pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "colorthief>=0.2.1",
    "discord-py>=2.5.0",
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "numpy>=2.2.3",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "trafilatura>=2.0.0",
    "webcolors>=24.11.1",
]
